requirements.txt
semantic-kernel==0.5.1.dev0
python-dotenv==1.0.0
openai==1.12.0
Faker==23.2.1
pyodbc==5.1.0
gradio==3.1.5
httpx==0.23.0
httpcore==0.15.0
